% Generated by roxygen2 (4.0.2): do not edit by hand
\name{daylights_filter}
\alias{daylights_filter}
\title{Function fixing daylights savings issues by replacing missing hours with mean
of neighbouring hours and collecting extra hour to single observation by
taking the mean.}
\usage{
daylights_filter(input_data)
}
\arguments{
\item{input_frame}{A dataframe containing data to be fixed.}
}
\value{
A dataframe, the original imported dataframe fixed.
}
\description{
Function fixing daylights savings issues by replacing missing hours with mean
of neighbouring hours and collecting extra hour to single observation by
taking the mean.
}

