% Generated by roxygen2 (4.0.2): do not edit by hand
\name{deseason_price}
\alias{deseason_price}
\title{Function for filtering seasonalities based on XXX}
\usage{
deseason_price(input_frame)
}
\arguments{
\item{input_frame}{A data frame containing the times series data.}
}
\value{
data_filt Same dataframe with deseasonalised price
}
\description{
Function for filtering seasonalities based on XXX
}

