% Generated by roxygen2 (4.0.2): do not edit by hand
\name{deseason_spot}
\alias{deseason_spot}
\title{Function for filtering seasonalities}
\usage{
deseason_spot(input_frame)
}
\arguments{
\item{input_frame}{A data frame containing the times series data.}
}
\value{
data_filt To be determined
}
\description{
Function for filtering seasonalities
}

