% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tests.R
\name{do_study}
\alias{do_study}
\title{Function performing all calculation and creating all tables for topic report}
\usage{
do_study(spot_data)
}
\description{
Function performing all calculation and creating all tables for topic report
}

