% Generated by roxygen2 (4.0.2): do not edit by hand
\name{draw_acf}
\alias{draw_acf}
\title{Flexible function for plotting autocorrelation functions in ggplot}
\usage{
draw_acf(input_frame, lags)
}
\arguments{
\item{input_frame}{A dataframe with a date object and a series to be plotted}

\item{lags}{An integer giving the number of lags}
}
\description{
Flexible function for plotting autocorrelation functions in ggplot
}

