% Generated by roxygen2 (4.0.2): do not edit by hand
\name{draw_periodogram}
\alias{draw_periodogram}
\title{Function plotting the periodogram for a univariate time series}
\usage{
draw_periodogram(input_frame)
}
\arguments{
\item{input_frame}{A dataframe with a series named price}
}
\description{
Function plotting the periodogram for a univariate time series
}

