% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_holi_dum}
\alias{get_holi_dum}
\title{Function getting holiday dummies}
\usage{
get_holi_dum(from_date, to_date, country = "de")
}
\arguments{
\item{from_date}{A string with the start date of the desired series}

\item{to_date}{A string with the end date of the desired series}

\item{country}{A string with the desired country iso id. Takes the values DE,
FR and CH. Default is DE.}
}
\value{
output_frame A dataframe containing date and seleced countries.
}
\description{
Function getting holiday dummies
}

