% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_treatment.R
\name{long_run_trend_season}
\alias{long_run_trend_season}
\title{Function for filtering seasonalities based on XXX}
\usage{
long_run_trend_season(input_frame)
}
\arguments{
\item{input_frame}{A data frame containing the times series data.}
}
\value{
data_filt Same dataframe with deseasonalised price
}
\description{
Function for filtering seasonalities based on XXX
}

