% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loss_function}
\alias{loss_function}
\title{Function for plotting the the epsilon-intensity linear loss function and
Huber's loss function.}
\usage{
loss_function(save_path = NULL, do_print = "Yes")
}
\arguments{
\item{save_path}{A path string. If given, saves an .eps file to local hard
drive}

\item{do_print}{A string, yes or no}
}
\description{
Function for plotting the the epsilon-intensity linear loss function and
Huber's loss function.
}

