% Generated by roxygen2 (4.0.2): do not edit by hand
\name{outlier_filt}
\alias{outlier_filt}
\title{Outlier filter function}
\usage{
outlier_filt(data_input, std_filt)
}
\arguments{
\item{data_input}{A data frame containing a price to be outlier filtered.}

\item{std_filt}{Number of standard deviations to filter on.}
}
\description{
Outlier filter function
}

