% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_treatment.R
\name{pre_model_processing}
\alias{pre_model_processing}
\title{Function carrying out the pre-processing of the data described in study data-
section. That is saving all plots and returning a filtered data_frame of the
entire period.}
\usage{
pre_model_processing(input_data, path_figure = NULL, path_table = NULL)
}
\arguments{
\item{path_figure}{A string with path to save figures}

\item{path_table}{A string with path to save tables}

\item{data_frame}{A data frame containing the spot price}
}
\description{
Function carrying out the pre-processing of the data described in study data-
section. That is saving all plots and returning a filtered data_frame of the
entire period.
}

