% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pre_model_processing_intraday}
\alias{pre_model_processing_intraday}
\title{Function to tread intraday spread data.}
\usage{
pre_model_processing_intraday(input_data, path_figure = NULL,
  path_table = NULL)
}
\description{
Function to tread intraday spread data.
}

