% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pre_model_processing_spot}
\alias{pre_model_processing_spot}
\title{Function carrying out the pre-processing of the data described in study data-
section. That is saving all plots and returning a filtered data_frame of the
entire period.}
\usage{
pre_model_processing_spot(input_data, path_figure = NULL, path_table = NULL)
}
\arguments{
\item{path_figure}{A string with path to save figures}

\item{path_table}{A string with path to save tables}

\item{data_frame}{A data frame containing the spot price}
}
\description{
Function carrying out the pre-processing of the data described in study data-
section. That is saving all plots and returning a filtered data_frame of the
entire period.
}

