% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scrape_epex_spot}
\alias{scrape_epex_spot}
\title{Function scraping various spotprices and volumes from the Epexspot website}
\usage{
scrape_epex_spot(from_date, to_date, country)
}
\arguments{
\item{from_date}{A string with the start date of the desired series}

\item{to_date}{A string with the end date of the desired series}

\item{country}{A string with the desired country iso id. Takes the values DE,
FR and CH.}
}
\value{
a dataframe containing the dates, hours, spot and volumes
}
\description{
Function scraping various spotprices and volumes from the Epexspot website
}

