% Generated by roxygen2 (4.0.2): do not edit by hand
\name{select_arima}
\alias{select_arima}
\title{Function doing base part of study}
\usage{
select_arima(input_frame, max.p = 7, max.q = 7)
}
\arguments{
\item{input_data}{A data frame}

\item{p}{An integer, number of max AR terms}

\item{q}{An integer, number of max MA terms}
}
\description{
Function doing base part of study
}

