% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model_building.R
\name{select_svm}
\alias{select_svm}
\title{Function basically performing the same task as e1071::tune, but here made
clear.}
\usage{
select_svm(input_frame, kernel = "linear", max.polynomial = 2,
  cachesize = 500, min_lag = 7, eps_grid = 10^c(0.5, 0.75, 1, 1.25, 1.5),
  gamma_grid = 10^c(0.5, 0.75, 1, 1.25, 1.5), sigma_grid = -10^seq(-1, 1,
  0.5), cores = 8L)
}
\arguments{
\item{input_frame}{A data frame containing a date and a price.}

\item{cachesize}{cache memory in MB (default 4000)}

\item{max.cost}{An integer containing the maximum allowed cost. If left as
NULL, it will default to floor(max(abs(price)) + 3 * sd(price))}

\item{max.epsilon}{An interger being the max allowed epsilon parameter. If
left as null, it will default to 4, being the value where no support vectors
are returned for the base spot price for Germany.}
}
\description{
Function basically performing the same task as e1071::tune, but here made
clear.
}

