% Generated by roxygen2 (4.0.2): do not edit by hand
\name{select_svm_bin}
\alias{select_svm_bin}
\title{Function to cross validate binary model}
\usage{
select_svm_bin(input_frame, kernel = "linear", max.polynomial = 4,
  cachesize = 8000, min_lag = 7)
}
\description{
Function to cross validate binary model
}

