% Generated by roxygen2 (4.0.2): do not edit by hand
\name{short_run_season}
\alias{short_run_season}
\title{Function calculating short run season}
\usage{
short_run_season(input_frame)
}
\arguments{
\item{input_frame}{A data frame with a trend and a price}
}
\description{
Function calculating short run season
}

