% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{svmfortopic}
\alias{svmfortopic}
\alias{svmfortopic-package}
\title{Package providing all data and programs used in the topic XXX by
Andreas Keller Leth Laursen}
\description{
Provides all data used in the project topic by Andreas Keller
Leth Laursen from public sources. Further all programs are included as well
as tables and graphs.
}
\details{
\tabular{ll}{
  Package: \tab svmfortopic \cr
  Type:    \tab Package     \cr
  Version: \tab 1.0         \cr
  Date:    \tab 2014-10-24  \cr
  License: \tab GPL         \cr
  }
}
\author{
Andreas Keller Leth Laursen
Maintainer Andreas Keller Leth Laursen <andreas.keller[at]gmail.com>
}

