% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tests.R
\name{do_unit_root_test}
\alias{do_unit_root_test}
\title{Function performing all calculation and creating all tables for topic report}
\usage{
do_unit_root_test(input_frame, save_path = NULL)
}
\description{
Function performing all calculation and creating all tables for topic report
}

