% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{draw_line_plot}
\alias{draw_line_plot}
\title{Flexible function for plotting line chart of data}
\usage{
draw_line_plot(input_frame, xlabel, ylabel, input, file_name = NULL,
  save_path = NULL, do_print = TRUE)
}
\arguments{
\item{input_frame}{A dataframe with a date object and a series to be plotted}

\item{xlabel}{A string containing x-label name}

\item{ylabel}{A string containing x-label name}
}
\description{
Flexible function for plotting line chart of data
}

