% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ewma}
\alias{ewma}
\title{Function calculating the exponentially weighted moving average of the input
time series.}
\usage{
ewma(input_vector, lambda = 0.975)
}
\arguments{
\item{input_vector}{An atomic vector containing a time series.}

\item{lambda}{A double, the lambda parameter for the ewma. Default value is
0.975}
}
\value{
An atomic vector with the transformed data.
}
\description{
Function calculating the exponentially weighted moving average of the input
time series.
}

