% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{loss_function}
\alias{loss_function}
\title{Function for plotting the the epsilon-intensity linear loss function and
Huber's loss function.}
\usage{
loss_function(save_path = NULL, do_print = TRUE)
}
\arguments{
\item{save_path}{A path string. If given, saves an .eps file to local hard
drive}

\item{do_print}{A string, yes or no}
}
\description{
Function for plotting the the epsilon-intensity linear loss function and
Huber's loss function.
}

