% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loss_functions}
\alias{loss_functions}
\title{Comput a number of statistical loss functions for forecasts}
\usage{
loss_functions(input_frame, table_path = NULL)
}
\description{
Comput a number of statistical loss functions for forecasts
}

