% Generated by roxygen2 (4.0.2): do not edit by hand
\name{max_marg_class}
\alias{max_marg_class}
\title{Function for plotting the maximal margin classifier}
\usage{
max_marg_class(set_seed = 1001, save_path = NULL, do_print = "Yes")
}
\arguments{
\item{set_seed}{An integer, setting the seed. Default value matches graph in
topic}

\item{save_path}{A path string. If given, saves an .eps file to local hard
drive}

\item{do_print}{A string, yes or no}
}
\description{
Function for plotting the maximal margin classifier
}

