% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{max_marg_class}
\alias{max_marg_class}
\title{Function for plotting the maximal margin classifier. Function inspired by
http://scikit-learn.org/0.10/auto_examples/svm/plot_separating_hyperplane.htm
l}
\usage{
max_marg_class(set_seed = 1001, save_path = NULL, do_print = TRUE)
}
\arguments{
\item{set_seed}{An integer, setting the seed. Default value matches graph in
topic}

\item{save_path}{A path string. If given, saves an .eps file to local hard
drive}

\item{do_print}{A string, yes or no}
}
\description{
Function for plotting the maximal margin classifier. Function inspired by
http://scikit-learn.org/0.10/auto_examples/svm/plot_separating_hyperplane.htm
l
}

