% Generated by roxygen2 (4.0.2): do not edit by hand
\name{na_filter}
\alias{na_filter}
\title{Function replacing missing values (NA) using linear interpolation between
closest finite observations. If starting points are missing, they are
replaced by the closets observation possible.}
\usage{
na_filter(input_data)
}
\arguments{
\item{input_data}{An atomic vector containing data to be fixed.}
}
\value{
An atomic vector filtered for missing values.
}
\description{
Function replacing missing values (NA) using linear interpolation between
closest finite observations. If starting points are missing, they are
replaced by the closets observation possible.
}

