% Generated by roxygen2 (4.0.2): do not edit by hand
\name{oos_forecast}
\alias{oos_forecast}
\title{Function doing oos forecast using a given model type over a given horison}
\usage{
oos_forecast(data_frame)
}
\description{
Function doing oos forecast using a given model type over a given horison
}

