% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model_building.R
\name{oos_forecast}
\alias{oos_forecast}
\title{Function doing oos forecast using a given model type over a given horison}
\usage{
oos_forecast(input_frame, test_start = "2013-11-01", p = 6, q = 5,
  h = 1)
}
\description{
Function doing oos forecast using a given model type over a given horison
}

