% Generated by roxygen2 (4.0.2): do not edit by hand
\name{outlier_filt}
\alias{outlier_filt}
\title{Outlier filter function}
\usage{
outlier_filt(input_frame, std_filt = 4)
}
\arguments{
\item{std_filt}{Number of standard deviations to filter on.}

\item{data_input}{A data frame containing a price to be outlier filtered.}
}
\description{
Outlier filter function
}

