% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_treatment.R
\name{outlier_filt}
\alias{outlier_filt}
\title{Outlier filter function}
\usage{
outlier_filt(input_frame, std_filt = 3)
}
\arguments{
\item{std_filt}{Number of standard deviations to filter on.}

\item{data_input}{A data frame containing a price to be outlier filtered.}
}
\description{
Outlier filter function
}

