% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_treatment.R
\name{pre_model_processing_intraday}
\alias{pre_model_processing_intraday}
\title{Function to tread intraday spread data.}
\usage{
pre_model_processing_intraday(input_data, path_figure = NULL,
  path_table = NULL)
}
\description{
Function to tread intraday spread data.
}

