% Generated by roxygen2 (4.0.2): do not edit by hand
\name{select_arima}
\alias{select_arima}
\title{Function doing base part of study}
\usage{
select_arima(input_frame, p = 7, q = 7)
}
\arguments{
\item{p}{An integer, number or AR terms}

\item{q}{An integer, number of MA terms}

\item{input_data}{A data frame}
}
\description{
Function doing base part of study
}

