% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model_building.R
\name{select_arima}
\alias{select_arima}
\title{Function doing base part of study}
\usage{
select_arima(input_frame, p = 7, q = 7)
}
\arguments{
\item{p}{An integer, number of max AR terms}

\item{q}{An integer, number of max MA terms}

\item{input_data}{A data frame}
}
\description{
Function doing base part of study
}

