% Generated by roxygen2 (4.0.2): do not edit by hand
\name{select_svm}
\alias{select_svm}
\title{Function basically performing the same task as e1071::tune, but here made
clear.}
\usage{
select_svm(input_frame, max.cost = NULL, max.epsilon = 4,
  cachesize = 4000)
}
\arguments{
\item{input_frame}{A data frame containing a date and a price.}

\item{max.cost}{An integer containing the maximum allowed cost. If left as
NULL, it will default to floor(max(abs(price)) + 3 * sd(price))}

\item{max.epsilon}{An interger being the max allowed epsilon parameter. If
left as null, it will default to 4, being the value where no support vectors
are returned for the base spot price for Germany.}

\item{cachesize}{cache memory in MB (default 4000)}
}
\description{
Function basically performing the same task as e1071::tune, but here made
clear.
}

