% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model_building.R
\name{select_svm_bin}
\alias{select_svm_bin}
\title{Function to cross validate binary model}
\usage{
select_svm_bin(input_frame, kernel = "linear", min.lag = 7,
  cost = 10^seq(-2, 1, 0.1), max.polynomial = 4, gamma = 10^seq(-6, 1,
  0.5), coef0 = -10^seq(-1, 1, 0.1), cachesize = 500, cores = 4L)
}
\description{
Function to cross validate binary model
}

