% Generated by roxygen2 (4.0.2): do not edit by hand
\name{select_svm_bin}
\alias{select_svm_bin}
\title{Function to cross validate binary model}
\usage{
select_svm_bin(input_frame, kernel = "linear", min.lag = 7,
  cost = 10^seq(-2, 1, 0.1), max.polynomial = 4, cachesize = 1000,
  cores = 4L)
}
\description{
Function to cross validate binary model
}

