% Generated by roxygen2 (4.0.2): do not edit by hand
\name{select_svm_bin}
\alias{select_svm_bin}
\title{Function to cross validate binary model}
\usage{
select_svm_bin(input_frame, kernel = "linear", min.lag = 7,
  cost = 10^seq(-2, 1, 0.1), max.polynomial = 4, gamma = 10^seq(-6, 1,
  0.5), coef0 = -10^seq(-1, 1, 0.1), cachesize = 500, cores = 8L)
}
\description{
Function to cross validate binary model
}

